Pong tutorial 06: Reduce the max framerate to 60 (important)
For unclear reasons, high framerates cause a jump in the frame time taken (that is, the allocated frame time, minus the sleep time).

For example, pong at 60 FPS takes a few ms per frame, but at 144 FPS takes more than 16 ms.

This has also been experienced with Dwarf seeks fortune; it seems to be relative to the computation demand, as with DSF, the jump happened already at around 60 FPS.

Tested on Windows.